refactor(attribution): neutralize acquisition attribution api naming
