+++ turbo/packages/api-contracts/src/contracts/index.ts
@@ -1243,12 +1243,12 @@ export {
 } from "./github-oauth";
 export {
   adAttributionMetadataSchema,
-  zeroAttributionContract,
+  acquisitionAttributionContract,
   type AdAttributionMetadata,
   type RecordSignupAttributionRequest,
   type RecordSignupAttributionResponse,
-  type ZeroAttributionContract,
-} from "./zero-attribution";
+  type AcquisitionAttributionContract,
+} from "./acquisition-attribution";
 export {
   zeroBillingStatusContract,
   zeroBillingCheckoutContract,
